requirements.txt
SpeechRecognition==3.8.1
python-dotenv==1.0.0
ibm-watson==7.0.0
ibm-cloud-sdk-core==3.16.7
google-generativeai==0.3.1
pynput==1.7.6
# PyAudio will be installed separately with special flags